refactor: structure application according to domains
