+++ src/main/java/com/example/petshop/config/mvc/MvcConfig.java
@@ -1,4 +1,4 @@
-package com.example.petshop;
+package com.example.petshop.config.mvc;
 
 import org.springframework.context.annotation.Configuration;
 import org.springframework.web.servlet.config.annotation.ViewControllerRegistry;
