+++ src/main/java/com/example/petshop/config/security/JwtToUserConverter.java
@@ -1,12 +1,12 @@
-package com.example.petshop.security;
+package com.example.petshop.config.security;
 
-import com.example.petshop.user.Role;
-import com.example.petshop.collection.User;
+import com.example.petshop.user.data.Role;
+import com.example.petshop.user.data.User;
+import org.springframework.core.convert.converter.Converter;
+import org.springframework.security.authentication.UsernamePasswordAuthenticationToken;
 import org.springframework.security.core.GrantedAuthority;
 import org.springframework.security.core.authority.SimpleGrantedAuthority;
 import org.springframework.security.oauth2.jwt.Jwt;
-import org.springframework.core.convert.converter.Converter;
-import org.springframework.security.authentication.UsernamePasswordAuthenticationToken;
 import org.springframework.stereotype.Component;
 
 import java.util.Collections;
