+++ src/main/java/com/example/petshop/config/security/KeyUtils.java
@@ -1,4 +1,4 @@
-package com.example.petshop.security;
+package com.example.petshop.config.security;
 
 import lombok.extern.slf4j.Slf4j;
 import org.springframework.beans.factory.annotation.Value;
