+++ src/main/java/com/example/petshop/config/security/TokenGenerator.java
@@ -1,7 +1,7 @@
-package com.example.petshop.security;
+package com.example.petshop.config.security;
 
-import com.example.petshop.collection.User;
-import com.example.petshop.dto.TokenDTO;
+import com.example.petshop.user.controller.dto.TokenDTO;
+import com.example.petshop.user.data.User;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.beans.factory.annotation.Qualifier;
 import org.springframework.security.authentication.BadCredentialsException;
