+++ src/main/java/com/example/petshop/pet/data/PetDataSeeder.java
@@ -1,18 +1,19 @@
-package com.example.petshop.seed;
+package com.example.petshop.pet.data;
 
-import com.example.petshop.dto.Type;
-import com.example.petshop.collection.Pet;
-import com.example.petshop.repository.PetRepository;
+import com.example.petshop.pet.data.Type;
+import com.example.petshop.pet.data.Pet;
+import com.example.petshop.pet.data.PetRepository;
 import net.datafaker.Faker;
 import org.bson.types.Binary;
 import org.springframework.beans.factory.annotation.Autowired;
+import org.springframework.boot.CommandLineRunner;
 import org.springframework.stereotype.Component;
 
 import java.util.List;
 import java.util.stream.IntStream;
 
 @Component
-public class PetDataSeeder {
+public class PetDataSeeder implements CommandLineRunner {
     private final PetRepository petRepository;
 
     @Autowired
@@ -37,6 +38,11 @@ public class PetDataSeeder {
                 faker.bool().bool(),
                 new Binary(faker.photography().camera().getBytes())
          );
+    }
+
+    @Override
+    public void run(String... args) throws Exception {
+        seedData();
     }
 }
