+++ src/main/java/com/example/petshop/user/controller/AuthController.java
@@ -1,11 +1,11 @@
-package com.example.petshop.controller;
+package com.example.petshop.user.controller;
 
-import com.example.petshop.collection.User;
-import com.example.petshop.dto.LoginDTO;
-import com.example.petshop.dto.SignupDTO;
-import com.example.petshop.dto.TokenDTO;
-import com.example.petshop.security.TokenGenerator;
-import com.example.petshop.service.UserService;
+import com.example.petshop.user.service.UserService;
+import com.example.petshop.user.controller.dto.LoginDTO;
+import com.example.petshop.user.controller.dto.SignupDTO;
+import com.example.petshop.user.controller.dto.TokenDTO;
+import com.example.petshop.config.security.TokenGenerator;
+import com.example.petshop.user.data.User;
 import org.springframework.beans.factory.annotation.Qualifier;
 import org.springframework.http.HttpStatus;
 import org.springframework.http.ResponseEntity;
@@ -14,10 +14,12 @@ import org.springframework.security.authentication.dao.DaoAuthenticationProvider
 import org.springframework.security.core.Authentication;
 import org.springframework.security.core.AuthenticationException;
 import org.springframework.security.core.authority.SimpleGrantedAuthority;
-import org.springframework.security.oauth2.jwt.Jwt;
 import org.springframework.security.oauth2.server.resource.BearerTokenAuthenticationToken;
 import org.springframework.security.oauth2.server.resource.authentication.JwtAuthenticationProvider;
-import org.springframework.web.bind.annotation.*;
+import org.springframework.web.bind.annotation.PostMapping;
+import org.springframework.web.bind.annotation.RequestBody;
+import org.springframework.web.bind.annotation.RequestMapping;
+import org.springframework.web.bind.annotation.RestController;
 import org.springframework.web.server.ResponseStatusException;
 
 import java.util.Collections;
@@ -43,7 +45,7 @@ public class AuthController {
     }
 
     @PostMapping("/register")
-    public ResponseEntity register(@RequestBody SignupDTO signupDTO) {
+    public ResponseEntity<TokenDTO> register(@RequestBody SignupDTO signupDTO) {
         try {
 
             String errors = UserService.validateUser(signupDTO);
@@ -61,23 +63,19 @@ public class AuthController {
             );
 
             return ResponseEntity.ok(tokenGenerator.createToken(authentication));
-        } catch (Exception ex) {
-            if (ex instanceof IllegalArgumentException) {
-
+        } catch (IllegalArgumentException ex) {
                 throw new ResponseStatusException(HttpStatus.BAD_REQUEST, ex.getMessage());
-            } else {
+        } catch (Exception ex) {
                 throw new ResponseStatusException(HttpStatus.INTERNAL_SERVER_ERROR, ex.getMessage());
-
             }
         }
-    }
   /*  @GetMapping("/register")
   public ModelAndView addLoginForm() {
      return new ModelAndView("login", Collections.singletonMap("signUpDTO", new SignupDTO()));
    }*/
 
     @PostMapping("/login")
-    public ResponseEntity login(@RequestBody LoginDTO loginDTO) {
+    public ResponseEntity<TokenDTO> login(@RequestBody LoginDTO loginDTO) {
         try {
             Authentication authentication = daoAuthenticationProvider.authenticate(
                     new UsernamePasswordAuthenticationToken(loginDTO.getEmail(),
@@ -91,7 +89,7 @@ public class AuthController {
     }
 
     @PostMapping("/token")
-    public ResponseEntity token(@RequestBody TokenDTO tokenDTO) {
+    public ResponseEntity<TokenDTO> token(@RequestBody TokenDTO tokenDTO) {
         Authentication authentication = refreshTokenAuthProvider.authenticate(new BearerTokenAuthenticationToken(tokenDTO.getRefreshToken()));
 
         return ResponseEntity.ok(tokenGenerator.createToken(authentication));
