+++ src/main/java/com/example/petshop/user/data/User.java
@@ -1,6 +1,5 @@
-package com.example.petshop.collection;
+package com.example.petshop.user.data;
 
-import com.example.petshop.user.Role;
 import com.example.petshop.serializer.ObjectIdSerializer;
 import com.fasterxml.jackson.databind.annotation.JsonSerialize;
 import lombok.AllArgsConstructor;
