+++ src/main/java/com/example/petshop/user/data/UserRepository.java
@@ -1,6 +1,5 @@
-package com.example.petshop.repository;
+package com.example.petshop.user.data;
 
-import com.example.petshop.collection.User;
 import org.bson.types.ObjectId;
 import org.springframework.data.mongodb.repository.MongoRepository;
 
@@ -9,4 +8,6 @@ import java.util.Optional;
 public interface UserRepository extends MongoRepository<User, ObjectId> {
    Optional<User> findByEmail(String email);
    boolean existsByEmail(String email);
+
+   void deleteByEmail(String username);
 }
